refactor: moved files into there respective packages for better usage
