+++ b3dm/b3dm.go
@@ -1,4 +1,4 @@
-package indexer
+package b3dm
 
 import (
 	"encoding/binary"
