--- indexer_config_test.go
@@ -1,15 +1,0 @@
-package indexer
-
-import (
-	"fmt"
-	"testing"
-)
-
-func TestParseIndexerConfig(t *testing.T) {
-	indexerConfig, err := ParseIndexerConfigFile("example/config.json")
-	fmt.Println(indexerConfig)
-
-	if err != nil {
-		t.Errorf("failed to parse the indexerconfig json")
-	}
-}
